Remove TestingBlobTableInfoFactory
We no longer use the ability to manually create `BlobTableInfo`
instances in the tests, but instead we use `addBlobTable` on the
`SQLExecutor.Builder` which adds the tables to the cluster state.
